+++ src/app/components/StandardSlot.jsx
@@ -46,7 +46,7 @@ export default class StandardSlot extends TranslatedComponent {
     let m = slot.m;
     let classRating = m.class + m.rating;
     let menu;
-    let validMods = m == null ? [] : (Modifications.validity[m.grp] || []);
+    let validMods = m == null ? [] : (Modifications.modules[m.grp].modifications || []);
     let showModuleResistances = Persist.showModuleResistances();
     let mass = m.getMass() || m.cargo || m.fuel || 0;