+++ src/app/shipyard/Ship.js
@@ -1335,7 +1335,7 @@ export default class Ship {
     if (this.bulkheads.m && this.bulkheads.m.mods) {
       for (let modKey in this.bulkheads.m.mods) {
         // Filter out invalid modifications
-        if (Modifications.validity['bh'] && Modifications.validity['bh'].indexOf(modKey) != -1) {
+        if (Modifications.modules['bh'] && Modifications.modules['bh'].modifications.indexOf(modKey) != -1) {
           bulkheadMods.push({ id: Modifications.modifications[modKey].id, value: this.bulkheads.m.getModValue(modKey) });
         }
       }
@@ -1350,7 +1350,7 @@ export default class Ship {
       if (slot.m && slot.m.mods) {
         for (let modKey in slot.m.mods) {
           // Filter out invalid modifications
-          if (Modifications.validity[slot.m.grp] && Modifications.validity[slot.m.grp].indexOf(modKey) != -1) {
+          if (Modifications.modules[slot.m.grp] && Modifications.modules[slot.m.grp].modifications.indexOf(modKey) != -1) {
             slotMods.push({ id: Modifications.modifications[modKey].id, value: slot.m.getModValue(modKey) });
           }
         }
@@ -1365,7 +1365,7 @@ export default class Ship {
       if (slot.m && slot.m.mods) {
         for (let modKey in slot.m.mods) {
           // Filter out invalid modifications
-          if (Modifications.validity[slot.m.grp] && Modifications.validity[slot.m.grp].indexOf(modKey) != -1) {
+          if (Modifications.modules[slot.m.grp] && Modifications.modules[slot.m.grp].modifications.indexOf(modKey) != -1) {
             slotMods.push({ id: Modifications.modifications[modKey].id, value: slot.m.getModValue(modKey) });
           }
         }
@@ -1380,7 +1380,7 @@ export default class Ship {
       if (slot.m && slot.m.mods) {
         for (let modKey in slot.m.mods) {
           // Filter out invalid modifications
-          if (Modifications.validity[slot.m.grp] && Modifications.validity[slot.m.grp].indexOf(modKey) != -1) {
+          if (Modifications.modules[slot.m.grp] && Modifications.modules[slot.m.grp].modifications.indexOf(modKey) != -1) {
             slotMods.push({ id: Modifications.modifications[modKey].id, value: slot.m.getModValue(modKey) });
           }
         }